pyproject.toml
[tool.poetry]
name = "attrs-marshmallow"
version = "0.6.1"
description = ""
authors = ["Igor Stuzhuk (KoHcoJlb) <fujitsuigor@gmail.com>"]
keywords = [
    "marshmallow",
    "attrs",
    "serialization",
    "deserialization"
]

[tool.poetry.urls]
Repository = "https://github.com/KoHcoJlb/attrs-marshmallow"

[tool.poetry.dependencies]
python = "^3.7"
marshmallow = ">=3.0"
attrs = "^19.1"
typing-inspect = "^0.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
